Clean up MaxLength extension
