+++ src/extensions/MaxLength/MaxLength.css
@@ -1,0 +1,3 @@
+.overflow-mark {
+  background-color: #ff4136;
+}
